Add role syncing module
This module will sync source and target roles. When a user gets added or removed the source role, the bot will copy those changes to the target role.

Roles can span across guilds. The `/sync-role` command can also be used to manually sync the roles.

The module is already configured to sync the "Quilt Developer" role from the Toolchain server to the Community server.