Showcase: ban some hostile reactions
This commit bans any reaction containing 'modrinth', 'curse'(forge) and 'github' in their name, to avoid constant pushing of developers to release to a new platform or open source their mod. I'm currently not happy with the DM message that I currently written, if anyone would like to submit a rewrite that'd be greatly appreciated.