Allow staff commands to community managers
Previously, the assumption was made that community managers would always have a moderator role, which doesn't hold up anymore. This change the hasBaseModeratorRole and notHasBaseModeratorRole checks to include community managers by default. We did not include that role for strictly moderation related commands: the filter and github extension.

Here is our reasoning for not blocking some commands that may be more obscure:
- Moderation extension: This extension currently only provides the /slowmode command, which allows for more freedom than the Discord client normally give you. Slowmode can be an integral part of setting up new channels or an organisation policy. We may need to introduce two checks in that extension in the long run.
- /lock and /unlock: We frequently lock and unlock channels outside of a moderation context, for example during meetings.

Feedbacks on what is allowed and not allowed is always welcome.